beancount_importers_india/BOI/requirements.txt
beancount
pandas
python-docx
loguru
beancount
docx2txt
